[refact] 주문 서비스 에러처리 고도화

+++ msa.order/src/main/java/com/msa/order/presentation/request/UpdateOrderRequest.java
@@ -16,6 +16,6 @@ public record UpdateOrderRequest(
     String itemName,
 
     @NotNull @Min(1)
-    int quantity) {
+    Long quantity) {
 
 }
